Store the IP address of the corresponding ifa in the rt_gateway field of RTF_CLONING and RTF_BROASCAST routes to not create MPATH conflicts when IP address aliases are used.
This change makes it possible to have multiple RTF_CLONING routes with
the same priority.  Note that any of the existing RTF_CLONING route
might be used by the kernel to create a RTF_CLONED route which should
not be a problem with aliases since they are attached to the same ifp.

This unbreak address aliases since the kernel supports multiple connected
routes for a subnet.

Found the hardway by djm@, ok claudio@