Do not trigger a KASSERT() if the route we're trying to remove does not exist and we get another matching one instead.
This bug has been here since the KAME area and recently exposed by a
refactoring at n2k15.  The problem is that rtrequest(9) does not check
on which interface the route entry is attached when issuing a RTM_DELETE.

So the kernel would end up deleting the route attached on a different ifp
when in_ifinit() fails.

This fix is currently a workaround, a better fix is in the pipeline.

Reported by Laurence Tratt <laurie AT tratt DOT net>, thanks!