Add pledge(2).  It is not currently possible to pledge -u and -U because of chown(2) and chmod(2).
If -u and -U aren't specified, the basic pledge is "stdio rpath getpw".
"getpw" is used to resolve uid and gid to names and vice versa.
The -r flag makes use of rmdir(2) and unlink(2), so we need "cpath"
The -t flag uses utimes(2), so we need "fattr"

ok deraadt@