Introduce xen_intr_mask and xen_intr_unmask primitives
Mask the event port during xen_intr_establish, but don't set the
masked flag in the intsrc.  By providing mask and unmask routines
we allow the device to decide when to perform these actions.  The
port will still be unmasked during xen_intr_enable.  This allows
netfront to fulfil the intr_barrier pattern requirements fairly
easily and at the same time should be sufficient for diskfront
that doesn't need to fiddle with interrupt masking.