UTF-8 support for fmt -c.
This implies two small changes in behaviour:
1. Let fmt -c replace invalid bytes with ASCII question marks
just like when called without -c.
2. On lines to be centered, replace each tab with a single blank,
simply because there is no useful way to define the meaning of a
tab on such a line.  Having the width of a tab depend on what is
to the right of it would be completely crazy (and complicate the
code a lot), and otherwise, tabs on adjacent lines of different
length wouldn't align anyway.

OK millert@