Next step in UTF-8 support:
1. Improve all functions involving words by allowing non-ASCII
characters to be part of words.
2. Allow insertion of non-ASCII characters without screwing up the
display, by backing up to the start byte after inserting a continuation
byte, and starting to re-print there.
3. Fix forward movement which i didn't get quite right in my previous
commit: Always advance to a start byte, never to a final continuation
byte, or the next insertion would split the character in the middle.

OK mpi@