add a calendar file for the uk;
this is the work of craig skinner - many thanks, craig.
i've tweaked it a bit. craig has also fixed a currently
incorrect date and moved a couple entries to the uk file/

some outstanding issues (what to do with dups) but it'll be
eaiser to work on once in tree