Order the routing instances by AF and then by AS.
This commit only improves the readability of the 'eigrpctl' show commands.

Additionaly, we can simplify the nbr_compare() and rt_compare() functions
because we know that under no circumstances nbrs or routes from different
AFs or ASes will be in the same tree. Each instance has its own trees of
neighbors and routes.