Do not set the EoT flag in the last startup update.
Unfortunately we don't have enough information to support the NSF feature,
the EIGRP draft is missing a lot of information in this regard.

The EoT flag is used as part of EIGRP NSF to announce the end of the
startup process with a neighbor. The problem is that, when we set this
flag, the Cisco neighbors will assume that we support NSF, which is
not true.  When this happens, these routers expect us to do things that
we don't know, and weird things can happen. So let's play safe and just
ignore this flag for now.